added a file exist check in the start script

+++ start.sh
@@ -2,6 +2,9 @@
 
 path=target/universal/stage
 
-sudo kill -SIGTERM `sudo cat $path/RUNNING_PID`
+if [ -f $path/RUNNING_PID ]; then
+    echo "Running pid found!"
+    sudo kill -SIGTERM `sudo cat $path/RUNNING_PID`
+fi
 
 sudo $path/bin/sproutup -Dconfig.file=conf/application-prod.conf -J-javaagent:/home/ubuntu/newrelic/newrelic.jar &
